Advanced-02-multimodal/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pygame>=2.0.0
